requirements.txt
opencv-python-headless==4.8.1.78
opencv-python==4.8.1.78
supervision==0.15.0
scikit-learn==1.1.2
ultralytics==8.0.221
mediapipe==0.10.8
streamlit==1.28.2
Pillow==10.1.0
torch==2.1.1
pandas==2.1.3